Split the viewer Qt code into multiple files
