+++ src/viewer/modmesh/viewer/RStaticMesh.hpp
@@ -43,11 +43,7 @@
 #include <Qt3DCore/QAttribute>
 #include <Qt3DCore/QTransform>
 
-#include <Qt3DRender/QCamera>
-
 #include <Qt3DExtras/QDiffuseSpecularMaterial>
-
-#include <QOrbitCameraController>
 
 namespace modmesh
 {
@@ -165,76 +161,6 @@ Qt3DRender::QGeometryRenderer * RStaticMesh<ND>::make_renderer(Qt3DCore::QGeomet
     return renderer;
 }
 
-class RScene
-    : public Qt3DCore::QEntity
-{
-
-public:
-
-    RScene(Qt3DCore::QNode * parent = nullptr)
-        : Qt3DCore::QEntity(parent)
-        , m_controller(new Qt3DExtras::QOrbitCameraController(this))
-    {
-    }
-
-    Qt3DExtras::QOrbitCameraController const * controller() const { return m_controller; }
-    Qt3DExtras::QOrbitCameraController * controller() { return m_controller; }
-
-private:
-
-    Qt3DExtras::QOrbitCameraController * m_controller = nullptr;
-
-}; /* end class RScene */
-
-class R3DWidget
-    : public QWidget
-{
-
-public:
-
-    R3DWidget(Qt3DExtras::Qt3DWindow * window = nullptr, RScene * scene = nullptr, QWidget * parent = nullptr, Qt::WindowFlags f = Qt::WindowFlags());
-
-    template <typename... Args>
-    void resize(Args &&... args);
-
-    Qt3DExtras::Qt3DWindow * view() { return m_view; }
-    RScene * scene() { return m_scene; }
-
-private:
-
-    Qt3DExtras::Qt3DWindow * m_view;
-    RScene * m_scene;
-    QWidget * m_container;
-
-}; /* end class R3DWidget */
-
-R3DWidget::R3DWidget(Qt3DExtras::Qt3DWindow * window, RScene * scene, QWidget * parent, Qt::WindowFlags f)
-    : QWidget(parent, f)
-    , m_view(nullptr == window ? new Qt3DExtras::Qt3DWindow : window)
-    , m_scene(nullptr == scene ? new RScene : scene)
-    , m_container(createWindowContainer(m_view, this, Qt::Widget))
-{
-    m_view->setRootEntity(m_scene);
-
-    // Set up the camera.
-    Qt3DRender::QCamera * camera = m_view->camera();
-    camera->lens()->setPerspectiveProjection(45.0f, 16.0f / 9.0f, 0.1f, 1000.0f);
-    camera->setPosition(QVector3D(0, 0, 40.0f));
-    camera->setViewCenter(QVector3D(0, 0, 0));
-
-    // Set up the camera control.
-    auto * control = m_scene->controller();
-    control->setCamera(m_view->camera());
-    control->setLinearSpeed(50.0f);
-    control->setLookSpeed(180.0f);
-}
-
-template <typename... Args>
-void R3DWidget::resize(Args &&... args)
-{
-    QWidget::resize(std::forward<Args>(args)...);
-    m_view->resize(std::forward<Args>(args)...);
-    m_container->resize(std::forward<Args>(args)...);
-}
-
 } /* end namespace modmesh */
+
+// vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
